style: make pre-commit pass on this branch
CI runs `pre-commit run --all-files` since #130. All 36 ruff findings were
in files this branch adds -- main itself is clean -- so they are ours to fix.

torch-calculate-electrostatic-potential/solvent:
- TID252: parent-relative imports (`from ..grid import`) made absolute
- TC002: `torch` moved into TYPE_CHECKING in occupancy.py and potential.py;
  neither uses it at runtime and both already have `from __future__ import
  annotations`
- RUF002/RUF003: en dashes and a Greek rho replaced with ASCII
- RUF046: redundant `int()` around `math.ceil`/`math.floor` in tests

Example notebooks:
- B905: `strict=True` on zips that are equal-length by construction
- RUF001/RUF003: ambiguous en dashes and minus signs in strings and comments
- D103: docstring for `sample_centers`
- E501: long lines wrapped in simulate_movies_to_tomogram.ipynb

The notebook edits are formatting and docstrings only, so the committed
outputs still correspond to the code that produced them.

_typos.toml gains `arange`, which is a numpy/torch API name rather than a
misspelling of "arrange". It only trips the hook inside notebooks, where
typos parses the JSON as prose instead of as code.

All five hooks pass (validate-pyproject, typos, ruff, ruff-format, mypy).
Suite: 1292 passed, 69 skipped.

+++ packages/primitives/torch-calculate-electrostatic-potential/tests/test_solvent.py
@@ -73,11 +73,11 @@ def test_occupancy_excludes_vdw_plus_probe_for_carbon():
     # A voxel just outside VdW+probe should be solvent.
     excluded_radius = vdw_c + probe_radius
     # Move ~ceil(excluded_radius)+1 voxels along x from center.
-    offset = int(math.ceil(excluded_radius)) + 1
+    offset = math.ceil(excluded_radius) + 1
     assert occupancy[center[0], center[1], center[2] + offset].item() == 1.0
 
     # A voxel inside the excluded ball (along axis, within floor(excluded_radius)-1).
-    inside = max(1, int(math.floor(excluded_radius)) - 1)
+    inside = max(1, math.floor(excluded_radius) - 1)
     assert occupancy[center[0], center[1], center[2] + inside].item() == 0.0
